Allow to use project for clang-tidy only
The requirements of cppcheck and clang-tidy are slightly conflicting.
Clang-tidy needs to know all the compiler options that are normally
needed for actual compilation of the analyzed source. Contrary to that,
when cppcheck is passed all these options (e.g. via
compile_commands.json), it ends up working with the actual system
headers, which brings unnecessary configuration variants to consider.